Split test_io.py into multiple files
